Tidy up the PR
