--- common/models/endpoint-group.json
@@ -1,28 +1,0 @@
-{
-  "name": "EndpointGroup",
-  "base": "Definition",
-  "idInjection": true,
-  "options": {
-    "validateUpsert": true
-  },
-  "properties": {
-    "id": {
-      "type": "string",
-      "id": true,
-      "json": false
-    },
-    "name": {
-      "type": "string",
-      "json": false
-    }
-  },
-  "validations": [],
-  "relations": {
-    "endpoints": {
-      "type": "referencesMany",
-      "model": "Endpoint"
-    }
-  },
-  "acls": [],
-  "methods": []
-}
