--- common/models/endpoint.js
@@ -1,3 +1,0 @@
-module.exports = function(Endpoint) {
-
-};
