--- common/models/endpoint.json
@@ -1,28 +1,0 @@
-{
-  "name": "Endpoint",
-  "base": "Definition",
-  "idInjection": true,
-  "options": {
-    "validateUpsert": true
-  },
-  "properties": {
-    "id": {
-      "type": "string",
-      "id": true,
-      "json": false
-    },
-    "name": {
-      "type": "string",
-      "json": false
-    },
-    "routes": {
-      "type": [
-        "string"
-      ]
-    }
-  },
-  "validations": [],
-  "relations": {},
-  "acls": [],
-  "methods": []
-}
