+++ common/models/reverse-proxy-policy.js
@@ -1,26 +1,12 @@
-var app;
-
 module.exports = function(ReverseProxy) {
-  app = ReverseProxy.app;
-  app.once('ready', function() {
-    ready(ReverseProxy);
-  });
-}
-
-function ready(ReverseProxy) {
-  var debug = require('debug')('workspace:policy:reverse-proxy');
-
-  var Middleware = app.models.Middleware;
-  var ReverseProxy = app.models.ReverseProxy;
-
   ReverseProxy.prototype.addMapping = function(mapping, cb) {
     this.params.routes = this.params.routes || [];
-    if(typeof mapping === 'string') {
-
-    } else if(typeof mapping === 'object') {
-
+    if (typeof mapping === 'string') {
+      // TBI
+    } else if (typeof mapping === 'object') {
+      // TBI
     } else {
-
+      // TBI
     }
   };
 };
