+++ package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "main": "server/server.js",
   "scripts": {
-    "pretest": "jshint --exclude-path .gitignore ."
+    "pretest": "jshint --exclude-path .gitignore .",
+    "test": "mocha -R spec"
   },
   "dependencies": {
     "async": "^1.3.0",
@@ -24,6 +25,8 @@
   },
   "devDependencies": {
     "chai": "^3.2.0",
+    "cookie-parser": "^1.3.5",
+    "fs-extra": "^0.22.1",
     "jshint": "^2.5.6",
     "loopback-workspace": "^3.15.1"
   },
