--- server/models/metrics-policy.js
@@ -1,5 +1,0 @@
-var app = require('../app');
-var debug = require('debug')('workspace:policy:metrics');
-
-var Metrics = app.models.Metrics;
-
