--- server/models/oauth2-policy.js
@@ -1,9 +1,0 @@
-var app = require('../app');
-var debug = require('debug')('workspace:policy:oauth2');
-
-var OAuth2 = app.models.OAuth2;
-
-OAuth2.prototype.defineScope = function(scope, cb) {
-
-};
-
