--- server/models/rate-limiting-policy.js
@@ -1,6 +1,0 @@
-var app = require('../app');
-var loopback = require('loopback');
-var debug = require('debug')('workspace:policy:rate-limiting');
-
-var RateLimiting = app.models.RateLimiting;
-
