+++ test/sandbox/server/config.json
@@ -1,0 +1,1 @@
+{}
