+++ test/sandbox/server/datasources.json
@@ -1,0 +1,1 @@
+{}
