+++ test/sandbox/server/model-config.json
@@ -1,0 +1,1 @@
+{}
